+++ src/gov/nih/nci/cbiit/scimgmt/gds/domain/RepositoryStatusesT.java
@@ -19,7 +19,7 @@ import javax.persistence.TemporalType;
 @Table(name = "REPOSITORY_STATUSES_T")
 public class RepositoryStatusesT implements java.io.Serializable {
 
-	private Integer id;
+	private Long id;
 	private LookupT lookupTByRepositoryId;
 	private LookupT lookupTByDataSubmissionStatusId;
 	private LookupT lookupTByRegistrationStatusId;
@@ -36,7 +36,7 @@ public class RepositoryStatusesT implements java.io.Serializable {
 	public RepositoryStatusesT() {
 	}
 
-	public RepositoryStatusesT(Integer id, LookupT lookupTByRepositoryId, LookupT lookupTByDataSubmissionStatusId,
+	public RepositoryStatusesT(Long id, LookupT lookupTByRepositoryId, LookupT lookupTByDataSubmissionStatusId,
 			LookupT lookupTByRegistrationStatusId, LookupT lookupTByStudyReleasedId, ProjectsT projectsT,
 			Date createdDate, String createdBy) {
 		this.id = id;
@@ -49,7 +49,7 @@ public class RepositoryStatusesT implements java.io.Serializable {
 		this.createdBy = createdBy;
 	}
 
-	public RepositoryStatusesT(Integer id, LookupT lookupTByRepositoryId, LookupT lookupTByDataSubmissionStatusId,
+	public RepositoryStatusesT(Long id, LookupT lookupTByRepositoryId, LookupT lookupTByDataSubmissionStatusId,
 			LookupT lookupTByRegistrationStatusId, LookupT lookupTByStudyReleasedId, ProjectsT projectsT,
 			Date anticipatedSubmissionDate, String accessionNumber, String comments, Date createdDate,
 			String createdBy, Date lastChangedDate, String lastChangedBy) {
@@ -71,11 +71,11 @@ public class RepositoryStatusesT implements java.io.Serializable {
 	@Id
 
 	@Column(name = "ID", unique = true, nullable = false, precision = 10, scale = 0)
-	public Integer getId() {
+	public Long getId() {
 		return this.id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
